feat: add leap year checker program
This commit adds a C program that reads a year from the user and displays a message indicating whether or not it is a leap year. According to the leap year rules, a year is a leap year if it is divisible by 4 but not by 100. However, years divisible by 400 are also considered leap years. The program uses if-else statements to make this determination.