git commit -m "Add C program to calculate factorial
Implemented a C program that calculates the factorial of a given number. The user is prompted to input a number, and the program validates the input. If the number is between 0 and 9 (inclusive), the factorial is computed using a loop. The result is then printed. Added input validation for a valid range.