Add C program to measure computer speed
This commit adds a C program that counts from 1 to 1,000,000,000 by 1s using a while loop. Every time the count reaches a multiple of 100,000,000, it prints that number on the screen. This program can be used to measure the speed of the computer by timing how long each 100 million repetitions of the loop takes.