Add program to sum a sequence of integers based on user input
This program reads an initial integer to determine the number of subsequent integers to be summed. It then prompts the user to enter each integer one by one, summing them as they are entered, and finally outputs the total sum. This approach follows the specific requirement that the first integer indicates the number of values to be summed and ensures that scanf is executed to read only one value each time.